requirements.txt
torch>=2.2.0
torchvision>=0.17.0
tenseal>=0.3.14
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=22.0.0
numpy>=1.26.0
requests>=2.31.0
Pillow>=10.2.0